+++ lib/src/classes/guild/unavailable_guild.dart
@@ -1,11 +1,9 @@
-import "guild.dart";
-
+/// A guild involved in a server outage.
 class UnavailableGuild {
+  /// The guild's id
   late String id;
-  late bool unavailable;
-  late Guild? notUpdatedGuild;
+  /// Whether the guild is available to access. If it is not available, it indicates a server outage
+  bool unavailable = true;
 
-  UnavailableGuild(this.id, { this.notUpdatedGuild }) {
-    unavailable = true;
-  }
+  UnavailableGuild(this.id);
 }