+++ lib/src/classes/images/base_image.dart
@@ -1,4 +1,0 @@
-class BaseImage {
-  late String hash;
-  BaseImage(this.hash);
-}